# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/users/{userID}/sites": {
            "get": {
                "description": "Retrieves sites for a specific user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user sites",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CapIot-api_internal_models.Site"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    }
                }
            }
        },
        "/api/assign-device": {
            "post": {
                "description": "Assigns a device to a location",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Assign device to location (Admin)",
                "parameters": [
                    {
                        "description": "Assignment details",
                        "name": "assignment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.AssignDeviceToLocationRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/devices-components-locations": {
            "get": {
                "description": "Retrieves a paginated list of devices, components, and locations for admin view",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get devices, components, and locations (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    }
                }
            }
        },
        "/api/location/{locationID}": {
            "put": {
                "description": "Modifies an existing location",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Modify location (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Location ID",
                        "name": "locationID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Location details",
                        "name": "location",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.Location"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a location",
                "tags": [
                    "admin"
                ],
                "summary": "Delete location (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Location ID",
                        "name": "locationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    }
                }
            }
        },
        "/api/locations-devices-users": {
            "get": {
                "description": "Retrieves a paginated list of locations, devices, and users for admin view",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get locations, devices, and users (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "CapIot-api_internal_models.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Use the new ErrorCode type",
                    "allOf": [
                        {
                            "$ref": "#/definitions/CapIot-api_internal_models.ErrorCode"
                        }
                    ]
                },
                "details": {
                    "description": "Optional: Additional details"
                },
                "message": {
                    "description": "Human-readable error message",
                    "type": "string"
                }
            }
        },
        "CapIot-api_internal_models.AssignDeviceToLocationRequest": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "location_id": {
                    "type": "integer"
                }
            }
        },
        "CapIot-api_internal_models.ErrorCode": {
            "type": "string",
            "enum": [
                "internal_server_error",
                "bad_request",
                "not_found",
                "forbidden",
                "unauthorized",
                "method_not_allowed",
                "invalid_token",
                "token_expired",
                "insufficient_permissions",
                "validation_failed",
                "missing_parameter",
                "invalid_format",
                "unsupported_media_type",
                "resource_not_found",
                "duplicate_resource"
            ],
            "x-enum-varnames": [
                "ErrorCodeInternalServerError",
                "ErrorCodeBadRequest",
                "ErrorCodeNotFound",
                "ErrorCodeForbidden",
                "ErrorCodeUnauthorized",
                "ErrorCodeMethodNotAllowed",
                "ErrorCodeInvalidToken",
                "ErrorCodeTokenExpired",
                "ErrorCodeInsufficientPermissions",
                "ErrorCodeValidationFailed",
                "ErrorCodeMissingParameter",
                "ErrorCodeInvalidFormat",
                "ErrorCodeUnsupportedMediaType",
                "ErrorCodeResourceNotFound",
                "ErrorCodeDuplicateResource"
            ]
        },
        "CapIot-api_internal_models.Location": {
            "type": "object",
            "properties": {
                "location_description": {
                    "type": "string"
                },
                "location_id": {
                    "type": "integer"
                },
                "location_name": {
                    "type": "string"
                },
                "site_id": {
                    "type": "integer"
                }
            }
        },
...
        "CapIot-api_internal_models.Site": {
            "type": "object",
            "properties": {
                "site_address": {
                    "type": "string"
                },
                "site_id": {
                    "type": "integer"
                },
                "site_name": {
                    "type": "string"
                }
            }
        },
...
    }
}